+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Fixes:
 - Update `munit` to 0.7.26 ([#75](https://github.com/MakeNowJust-Labo/recheck/pull/75))
 - Update `sbt-scoverage` to 1.8.0 ([#76](https://github.com/MakeNowJust-Labo/recheck/pull/76))
 - Update Scala version to 2.13.6
+- Run `recheck batch` requests in asynchronous ([#89](https://github.com/MakeNowJust-Labo/recheck/pull/89))
 
 # 3.1.0 (2021-05-14)